% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_system.R
\name{apply_system}
\alias{apply_system}
\title{Apply Integrated EPC-SPC to Real Data}
\usage{
apply_system(real_data, calibration, spc_type = "ewma", lambda = 0.1)
}
\arguments{
\item{real_data}{Numeric vector of actual, uncontrolled process readings}

\item{calibration}{A list object returned by \code{calibrate_system()}}

\item{spc_type}{Type of chart: "shewhart", "ewma", or "cusum" (default "ewma")}

\item{lambda}{EWMA weight parameter (default 0.1)}
}
\value{
A data frame of recommended adjustments and SPC signals
}
\description{
Apply Integrated EPC-SPC to Real Data
}
