% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calc_pm}
\alias{calc_pm}
\title{Calculate Performance Measure (PM)}
\usage{
calc_pm(y, target = 0)
}
\arguments{
\item{y}{Numeric vector of process outputs}

\item{target}{Numeric target value (default 0)}
}
\value{
A numeric scalar representing the PM
}
\description{
Calculate Performance Measure (PM)
}
