% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator.R
\name{evaluate_system}
\alias{evaluate_system}
\title{Evaluate SPC-EPC System Performance}
\usage{
evaluate_system(
  iterations = 100,
  n = 500,
  shift_time = 251,
  disturbance_type = "shift",
  shift_mag = 10,
  trend_mag = 0.1,
  spc_type = "shewhart",
  intervention = TRUE
)
}
\arguments{
\item{iterations}{Number of simulation runs (default 100)}

\item{n}{Observations per run (default 500)}

\item{shift_time}{Time the assignable cause occurs (default 251)}

\item{disturbance_type}{"shift" or "trend"}

\item{shift_mag}{Magnitude of the shift}

\item{trend_mag}{Magnitude of the trend}

\item{spc_type}{"shewhart", "ewma", or "cusum"}

\item{intervention}{Logical: Does the operator fix the process?}
}
\value{
A list containing average PM and average ARL
}
\description{
Evaluate SPC-EPC System Performance
}
