% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_control_action.R
\name{plot_control_action}
\alias{plot_control_action}
\title{Plot EPC Control Action}
\usage{
plot_control_action(sim_data)
}
\arguments{
\item{sim_data}{A data frame generated by simulate_epc_spc() or apply_system()}
}
\value{
A ggplot object
}
\description{
Visualizes the physical adjustments made by the automated controller over time.
}
