% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cusum.R
\name{plot_cusum}
\alias{plot_cusum}
\title{Plot CUSUM Chart}
\usage{
plot_cusum(sim_data)
}
\arguments{
\item{sim_data}{A data frame generated by simulate_epc_spc(spc_type = "cusum")}
}
\value{
A ggplot object
}
\description{
Plot CUSUM Chart
}
