% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ewma.R
\name{plot_ewma}
\alias{plot_ewma}
\title{Plot EWMA Chart}
\usage{
plot_ewma(sim_data)
}
\arguments{
\item{sim_data}{A data frame generated by simulate_epc_spc() or apply_system()}
}
\value{
A ggplot object
}
\description{
Plot EWMA Chart
}
