% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spc.R
\name{plot_spc}
\alias{plot_spc}
\title{Plot Integrated EPC-SPC Data}
\usage{
plot_spc(sim_data)

plot_spc(sim_data)
}
\arguments{
\item{sim_data}{A data frame generated by simulate_epc_spc()}
}
\value{
A ggplot object displaying the Shewhart chart

A ggplot object displaying the Shewhart chart
}
\description{
Plot Integrated EPC-SPC Data

Plot Integrated EPC-SPC Data
}
